feat(cmd): Add web target
Add web target.
