+++ api/web/api.go
@@ -1,0 +1,16 @@
+package web
+
+import (
+	"github.com/gin-gonic/gin"
+	"net/http"
+)
+
+func ApiRoot(c *gin.Context) {
+	c.String(http.StatusOK, "Hello gin")
+}
+
+func ApiConfigs(c *gin.Context) {
+	c.JSON(http.StatusOK, gin.H{
+		"configs": getConfigs(),
+	})
+}
